+++ tsconfig.json
@@ -1,0 +1,27 @@
+{
+    "compilerOptions": {
+        "outDir": ".",
+        "module": "commonjs",
+        "target": "esnext",
+        "lib": ["esnext"],
+        "declaration": true,
+        "strict": true,
+        "noImplicitAny": true,
+        "strictNullChecks": true,
+        "strictFunctionTypes": true,
+        "strictBindCallApply": true,
+        "strictPropertyInitialization": true,
+        "noImplicitThis": true,
+        "alwaysStrict": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "noImplicitReturns": true,
+        "noFallthroughCasesInSwitch": true,
+        "esModuleInterop": true,
+        "preserveConstEnums": true,
+        "moduleResolution": "node",
+        "types": ["@cloudflare/workers-types"]
+    },
+    "include": ["src"],
+    "exclude": ["node_modules"]
+}
